+++ demo/src/client/faro/initialize.ts
@@ -3,7 +3,6 @@ import { createRoutesFromChildren, matchRoutes, Routes, useLocation, useNavigati
 import {
   initializeFaro as coreInit,
   getWebInstrumentations,
-  InternalLoggerLevel,
   ReactIntegration,
   ReactRouterVersion,
 } from '@grafana/faro-react';