+++ packages/web-sdk/src/instrumentations/performance/types.ts
@@ -2,6 +2,7 @@ export type FaroNavigationTiming = Readonly<
   {
     duration: string;
     visibilityState: DocumentVisibilityState;
+    documentParsingTime: string;
     domProcessingTime: string;
     pageLoadTime: string;
     domContentLoadHandlerTime: string;
@@ -18,6 +19,7 @@ export type FaroResourceTiming = Readonly<{
   tcpHandshakeTime: string;
   dnsLookupTime: string;
   tlsNegotiationTime: string;
+  responseStatus: string;
   redirectTime: string;
   requestTime: string;
   fetchTime: string;